+++ aiida-registry-app/src/App.css
@@ -70,7 +70,6 @@ h1 {
 #app-body {
   max-width: 1280px;
   margin: 0 auto;
-  padding: 2rem;
 }
 
 .plugin-header, .plugin-header a {
